app/core/security.py
from passlib.context import CryptContext
from datetime import datetime, timedelta
from app.core.config import ACCESS_TOKEN_EXPIRE_MINUTES, SECRET_KEY,ALGORITMH
from jose import jwt, JWTError
from fastapi.security import OAuth2PasswordBearer
from fastapi import Depends, HTTPException, status
from app.db.session import get_db
from sqlalchemy.orm import Session
from app.models.usuarios import Usuario

pwd_context=CryptContext(
    schemes =["bcrypt"],
    deprecated="auto"
)
 

def hash_password(password:str)->str:
    return pwd_context.hash(password)

def verify_password(plain_password:str, hashed_password:str)->bool:
    return pwd_context.verify(plain_password, hashed_password)

def create_access_token(data:dict, expires_delta:timedelta|None=None):
    to_encode=data.copy()

    expire=datetime.utcnow()+(
        expires_delta
        if expires_delta
        else
        timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    )

    to_encode.update({"exp":expire})

    encoded_jwt=jwt.encode(
        to_encode,
        SECRET_KEY,
        algorithm=ALGORITMH
    )

    return encoded_jwt

oauth2_scheme=OAuth2PasswordBearer(tokenUrl="login")

def get_current_user(
        token:str=Depends(oauth2_scheme),
        db:Session=Depends(get_db)
):
    credentials_exception=HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="No autenticado",
        headers={"WWW-Authenticate":"Bearer"}
    )

    try:
        payload=jwt.decode(
            token,
            SECRET_KEY,
            algorithms=[ALGORITMH]
        )
        user_id:str=payload.get("sub")
        if user_id is None:
            raise credentials_exception
        
    except JWTError:
        raise credentials_exception
    
    user=db.query(Usuario).filter(Usuario.id==int(user_id)).first()

    if user is None:
        raise credentials_exception
    
    return user