requirements.txt
uvicorn 
fastapi
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
python-dotenv